EndpointsTestSuite and EndpointsTest extracted
