+++ akka-http/server-playjson/src/test/scala/endpoints/akkahttp/server/playjson/ServerInterpreterBaseTest.scala
@@ -10,6 +10,7 @@ import endpoints.algebra.server.{DecodedUrl, ServerTestBase}
 
 import scala.concurrent.duration._
 import scala.concurrent.{Await, Future}
+import endpoints.akkahttp.server.EndpointsTestApi 
 
 class ServerInterpreterBaseTest
   extends ServerTestBase[EndpointsTestApi]
