+++ akka-http/server-playjson/src/test/scala/endpoints/akkahttp/server/playjson/ServerInterpreterTest.scala
@@ -2,6 +2,7 @@ package endpoints.akkahttp.server.playjson
 
 import endpoints.akkahttp.server.EndpointsTestSuite
 import org.scalatest.{Matchers, WordSpecLike}
+import endpoints.akkahttp.server.EndpointsTestApi 
 
 class ServerInterpreterTest extends ServerInterpreterBaseTest
   with EndpointsTestSuite[EndpointsTestApi]
