+++ akka-http/server/src/test/scala/endpoints/akkahttp/server/EndpointsTest.scala
@@ -1,10 +1,9 @@
-package endpoints.akkahttp.server.playjson
+package endpoints.akkahttp.server
 
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import endpoints.akkahttp.server.{BasicAuthentication, Endpoints, JsonEntities, JsonEntitiesFromCodec}
+import endpoints.akkahttp.server.{BasicAuthentication, Endpoints, JsonEntities}
 import endpoints.algebra
-import endpoints.algebra.{JsonFromCodecTestApi, playjson}
 import org.scalatest.{Matchers, WordSpec}
 
 import scala.language.reflectiveCalls
@@ -14,13 +13,6 @@ trait EndpointsTestApi extends Endpoints
   with BasicAuthentication
   with algebra.BasicAuthTestApi
   with algebra.EndpointsTestApi
-
-/* implements the endpoint using a codecs-based json handling */
-class EndpointsCodecsTestApi extends EndpointsTestApi
-  with JsonFromCodecTestApi
-  with playjson.JsonFromPlayJsonCodecTestApi
-  with JsonEntitiesFromCodec
-  with playjson.JsonEntitiesFromCodec
 
 /* implements the endpoint using an akka-based custom json handling */
 class EndpointsEntitiesTestApi extends EndpointsTestApi
